binance_data_collector/requirements.txt
python-binance
requests
pandas
websocket-client
pyyaml 
binance-futures-connector